Move entity pagination tests to a different file
